Refactor: Move logic from UserPropertyAuthorizationRepository to a base class
This refactor will ease the implementation of non-updateable auth repos
