--- core/core/src/main/java/org/visallo/core/model/user/UpdatablePrivilegeRepository.java
@@ -1,9 +1,0 @@
-package org.visallo.core.model.user;
-
-import org.visallo.core.user.User;
-
-import java.util.Set;
-
-public interface UpdatablePrivilegeRepository extends PrivilegeRepository {
-    void setRoles(User user, Set<String> roles, User authUser);
-}
